+++ resources/js/components/Requirement/InstallForm.vue
@@ -53,6 +53,8 @@
                         type: 'RequirementInstall',
                         requirement: this.requirement,
                     });
+
+                    await this.$store.dispatch('autoRefreshJobs');
                 } catch (exception) {
                     this.alertError(exception);
                 }
@@ -61,7 +63,7 @@
             },
 
             isDisabled() {
-                return this.$store.state.isComposerBusy
+                return this.$store.state.isInstallerLocked
                     || this.$store.getters.getActiveJobs().length > 0;
             },