+++ resources/js/screens/requirements/index.vue
@@ -76,16 +76,7 @@
             Suggestions,
         },
 
-        mounted() {
-            this.fetchData();
-        },
-
         methods: {
-            async fetchData() {
-                await this.$store.dispatch('collectRequirements');
-                await this.$store.dispatch('collectJobs');
-            },
-
             getRequirements(level, environment) {
                 const requirements = this.$store.state.requirements;